Port AStar tests to use doctest
